requirements.txt
dask==2023.12.1
pandas==2.0.1
python-dotenv==1.0.0
tqdm==4.65.0
scipy==1.10.1
pyarrow==15.0.0
psutil==5.9.5
streamlit==1.28.1
stqdm==0.0.5